Refactored Keyboard Input
I refactored how the AdventureGame listens to the player input.

Created PlayerController.cs to listen to all input from the keyboard.

Tried to move some code around so that certain functions exist where they make the most sense to me.

The Player class has a PlayerController attached to it. The Player knows what Room (State) it is in. The Room knows what other Rooms are attached to it.